Complete OrderTable separate component
